% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakPC2017City10.R
\docType{data}
\name{PakPC2017City10}
\alias{PakPC2017City10}
\title{Top 10 Cities data from Pakistan Population Census 2017}
\format{A \code{data.table} and  \code{data.frame} with 10 obs. of  3 variables.}
\usage{
data(PakPC2017City10)
}
\description{
\code{PakPC2017City10} Top 10 Cities data from Pakistan Population Census 2017.
}
\details{
\itemize{
       \item City Cities of Pakistan
       \item Pop2017 Total Population in 2017
       \item Pop1998 Total Population in 1998
       }
}
\examples{
data(PakPC2017City10)
}
\references{
\enumerate{
\item Pakistan Population Census 2017 (\url{http://www.pbscensus.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2017Balochistan}}
 , \code{\link{PakPC2017FATA}}
 , \code{\link{PakPC2017Islamabad}}
 , \code{\link{PakPC2017KPK}}
 , \code{\link{PakPC2017Pak}}
 , \code{\link{PakPC2017Pakistan}}
 , \code{\link{PakPC2017Punjab}}
 , \code{\link{PakPC2017Sindh}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
